Adding FASTA files
I'm looking through the Google Doc and a lot of the files we've linked
so far aren't necessarily *THE* rhodopsin gene; a lot of them are
"partials" (e.g. "exon 1 through exon 4" and not the whole gene) or mRNA
(a different type of biological strand of coding information) rather
than the proper DNA sequence that we need. Or they're entire
chromosomes.

I have only been looking through the NIH website so maybe there are
others? If not, we may need to either cut out many species or keep
trying to find the relevant rhosopsin files for new species.